refactor(vs-ai): move headers into public include tree
