+++ app/src/main/cpp/PvZ/include/PvZ/Lawn/Board/VSActionAI/VSActionAIPlant/PlantAIPlanning.h
@@ -20,8 +20,8 @@
 #ifndef PVZ_LAWN_BOARD_VS_ACTION_AI_PLANT_AI_PLANNING_H
 #define PVZ_LAWN_BOARD_VS_ACTION_AI_PLANT_AI_PLANNING_H
 
-#include "Lawn/Board/VSActionAI/VSActionAIStrategy.h"
-#include "PlantDecisionContext.h"
+#include "PvZ/Lawn/Board/VSActionAI/VSActionAIStrategy.h"
+#include "PvZ/Lawn/Board/VSActionAI/VSActionAIPlant/PlantDecisionContext.h"
 
 #include <limits>
 #include <optional>
