+++ app/src/main/cpp/PvZ/include/PvZ/Lawn/Board/VSActionAI/VSActionAITacticalRules.h
@@ -20,7 +20,7 @@
 #ifndef PVZ_LAWN_BOARD_VS_ACTION_AI_TACTICAL_RULES_H
 #define PVZ_LAWN_BOARD_VS_ACTION_AI_TACTICAL_RULES_H
 
-#include "VSActionAIThreat.h"
+#include "PvZ/Lawn/Board/VSActionAI/VSActionAIThreat.h"
 
 namespace vsai::detail {
