+++ app/src/main/cpp/PvZ/include/PvZ/Lawn/Board/VSActionAI/VSActionAIZombie/ZombieCardRules.h
@@ -20,7 +20,7 @@
 #ifndef PVZ_LAWN_BOARD_VS_ACTION_AI_ZOMBIE_CARD_RULES_H
 #define PVZ_LAWN_BOARD_VS_ACTION_AI_ZOMBIE_CARD_RULES_H
 
-#include "Lawn/Board/VSActionAI/VSActionAIStrategy.h"
+#include "PvZ/Lawn/Board/VSActionAI/VSActionAIStrategy.h"
 
 #include <cstdint>
