+++ app/src/main/cpp/PvZ/src/Lawn/Board/VSActionAI/VSActionAIPlant/PlantEmergencyPolicy.cpp
@@ -17,7 +17,7 @@
  * PlantsVsZombies-AndroidTV. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "PlantAI.h"
+#include "PvZ/Lawn/Board/VSActionAI/VSActionAIPlant/PlantAI.h"
 
 #include <algorithm>
